+++ src/main/java/fr/oxal/v2/waven/entity/pvm/equipment/gem/Gem.java
@@ -24,7 +24,7 @@ public class Gem extends Equipment implements WithEffect, WithSpells {
     }
 
     @Override
-    public List<Double> getIdSpells() {
+    public List<Integer> getIdSpells() {
         return getIdSpells(getJsonRepresentation());
     }
